@dranikpg 2023-03-25
Fixed.

> My only concern is around Apply() bottleneck

Yes, I didn't yet optimize it - we should squash parallel updates in the future.